trafffic-mcp/requirements.txt
mcp
pandas
openpyxlmatplotlib